// interfaces/product.interface.ts
export interface ProductModel {
	_id: string
	categories: string[]
	tags: string[]
	title: string
	image: string
	description: string
	link: string
	price: number
	credit: number
	oldPrice: number
	characteristics: ProductCharacteristic[]
	advantages?: string
	initialRating: number
	createdAt: string
	updatedAt: string
	__v: number
	html: string
	blog: ProductBlog
	companyId: string
	clicks: number
	reviews: ProductReview[]
	reviewCount: number
	reviewAvg?: number
}

export interface ProductReview {
	_id: string
	name: string
	title: string
	description: string
	rating: number
	productId: string
	createdAt: string
	updatedAt: string
	__v: number
}

export interface ProductBlog {
	text: string
	_id: string
	bigImage?: string
}

export interface ProductCharacteristic {
	name: string
	value: string
}